Don't override http.Client checkRedirect on go1.8 or later
checkRedirect will replay headers on go1.8 or later so we don't need to
override it.  By not overriding it, we can avoid errors when running with the
race condition check.
