+++ check_redirect_go1.8.go
@@ -1,0 +1,9 @@
+// +build go1.8
+
+package eventsource
+
+import "net/http"
+
+// On go1.8 we don't need to do anything because it will copy headers on redirects
+func setCheckRedirect(c *http.Client) {
+}
